server/requirements.txt
Flask
requests
oxenai
